Read loop now empties input packet buffer each time it wakes up.
Earlier implementation was that the read loop wakes up and checks if there are any
packets in the input buffer. If so it processed only 1 packet per iteration. This commit
changes it so that each time it wakes up and finds that there are packets in the input
buffer, it will process all the packets until the input buffer is empty. This results in
faster behavior of packet handling and lesser dropped packets